Add left hand nav tabs
