+++ mlflow/server/js/src/gateway/components/edit-endpoint/index.ts
@@ -1,0 +1,2 @@
+export { EditEndpointFormRenderer } from './EditEndpointFormRenderer';
+export type { EditEndpointFormRendererProps } from './EditEndpointFormRenderer';
