+++ mlflow/server/js/src/gateway/pages/GatewayPage.tsx
@@ -1,6 +1,5 @@
-import { ScrollablePageWrapper } from '../../common/components/ScrollablePageWrapper';
 import { useEndpointsQuery } from '../hooks/useEndpointsQuery';
-import { Alert, Button, Header, PlusIcon, Spacer, useDesignSystemTheme } from '@databricks/design-system';
+import { Alert, Button, PlusIcon, Spacer, useDesignSystemTheme } from '@databricks/design-system';
 import { FormattedMessage } from 'react-intl';
 import { withErrorBoundary } from '../../common/utils/withErrorBoundary';
 import ErrorUtils from '../../common/utils/ErrorUtils';
@@ -13,33 +12,30 @@ const GatewayPage = () => {
   const { error, refetch } = useEndpointsQuery();
 
   return (
-    <ScrollablePageWrapper css={{ overflow: 'hidden', display: 'flex', flexDirection: 'column' }}>
-      <Spacer shrinks={false} />
-      <Header
-        title={
-          <FormattedMessage defaultMessage="Gateway" description="Header title for the gateway configuration page" />
-        }
-        buttons={
-          <Link to={GatewayRoutes.createEndpointPageRoute}>
-            <Button componentId="mlflow.gateway.create-endpoint-button" type="primary" icon={<PlusIcon />}>
-              <FormattedMessage defaultMessage="Create endpoint" description="Button to create endpoint" />
-            </Button>
-          </Link>
-        }
-      />
-      <Spacer shrinks={false} />
-      <div css={{ flex: 1, display: 'flex', flexDirection: 'column', overflow: 'hidden' }}>
-        {error && (
-          <>
-            <Alert type="error" message={error.message} componentId="mlflow.gateway.error" closable={false} />
-            <Spacer />
-          </>
-        )}
-        <div css={{ padding: theme.spacing.md, flex: 1, overflow: 'auto' }}>
-          <EndpointsList onEndpointDeleted={() => refetch()} />
-        </div>
+    <div css={{ flex: 1, display: 'flex', flexDirection: 'column', overflow: 'hidden' }}>
+      <div
+        css={{
+          display: 'flex',
+          justifyContent: 'flex-end',
+          padding: `${theme.spacing.sm}px ${theme.spacing.md}px`,
+          borderBottom: `1px solid ${theme.colors.border}`,
+        }}
+      >
+        <Link to={GatewayRoutes.createEndpointPageRoute}>
+          <Button componentId="mlflow.gateway.create-endpoint-button" type="primary" icon={<PlusIcon />}>
+            <FormattedMessage defaultMessage="Create endpoint" description="Button to create endpoint" />
+          </Button>
+        </Link>
       </div>
-    </ScrollablePageWrapper>
+      {error && (
+        <div css={{ padding: theme.spacing.md }}>
+          <Alert type="error" message={error.message} componentId="mlflow.gateway.error" closable={false} />
+        </div>
+      )}
+      <div css={{ padding: theme.spacing.md, flex: 1, overflow: 'auto' }}>
+        <EndpointsList onEndpointDeleted={() => refetch()} />
+      </div>
+    </div>
   );
 };
