+++ mlflow/server/js/src/gateway/route-defs.ts
@@ -5,23 +5,35 @@ export const getGatewayRouteDefs = () => {
   return [
     {
       path: GatewayRoutePaths.gatewayPage,
-      element: createLazyRouteElement(() => import('./pages/GatewayPage')),
+      element: createLazyRouteElement(() => import('./pages/GatewayLayout')),
       pageId: GatewayPageId.gatewayPage,
-    },
-    {
-      path: GatewayRoutePaths.createEndpointPage,
-      element: createLazyRouteElement(() => import('./pages/CreateEndpointPage')),
-      pageId: GatewayPageId.createEndpointPage,
-    },
-    {
-      path: GatewayRoutePaths.endpointDetailsPage,
-      element: createLazyRouteElement(() => import('./pages/EndpointDetailsPage')),
-      pageId: GatewayPageId.endpointDetailsPage,
-    },
-    {
-      path: GatewayRoutePaths.editEndpointPage,
-      element: createLazyRouteElement(() => import('./pages/EditEndpointPage')),
-      pageId: GatewayPageId.editEndpointPage,
+      children: [
+        {
+          index: true,
+          element: createLazyRouteElement(() => import('./pages/GatewayPage')),
+          pageId: GatewayPageId.gatewayPage,
+        },
+        {
+          path: 'api-keys',
+          element: createLazyRouteElement(() => import('./pages/ApiKeysPage')),
+          pageId: GatewayPageId.apiKeysPage,
+        },
+        {
+          path: 'endpoints/create',
+          element: createLazyRouteElement(() => import('./pages/CreateEndpointPage')),
+          pageId: GatewayPageId.createEndpointPage,
+        },
+        {
+          path: 'endpoints/:endpointId',
+          element: createLazyRouteElement(() => import('./pages/EndpointDetailsPage')),
+          pageId: GatewayPageId.endpointDetailsPage,
+        },
+        {
+          path: 'endpoints/:endpointId/edit',
+          element: createLazyRouteElement(() => import('./pages/EditEndpointPage')),
+          pageId: GatewayPageId.editEndpointPage,
+        },
+      ],
     },
   ];
 };
