+++ mlflow/server/js/src/gateway/routes.ts
@@ -2,6 +2,7 @@ import { createMLflowRoutePath, generatePath } from '../common/utils/RoutingUtil
 
 export enum GatewayPageId {
   gatewayPage = 'mlflow.gateway',
+  apiKeysPage = 'mlflow.gateway.api-keys',
   createEndpointPage = 'mlflow.gateway.create',
   endpointDetailsPage = 'mlflow.gateway.details',
   editEndpointPage = 'mlflow.gateway.edit',
@@ -12,22 +13,30 @@ export class GatewayRoutePaths {
     return createMLflowRoutePath('/gateway');
   }
 
+  static get apiKeysPage() {
+    return createMLflowRoutePath('/gateway/api-keys');
+  }
+
   static get createEndpointPage() {
-    return createMLflowRoutePath('/gateway/create');
+    return createMLflowRoutePath('/gateway/endpoints/create');
   }
 
   static get endpointDetailsPage() {
-    return createMLflowRoutePath('/gateway/:endpointId');
+    return createMLflowRoutePath('/gateway/endpoints/:endpointId');
   }
 
   static get editEndpointPage() {
-    return createMLflowRoutePath('/gateway/:endpointId/edit');
+    return createMLflowRoutePath('/gateway/endpoints/:endpointId/edit');
   }
 }
 
 class GatewayRoutes {
   static get gatewayPageRoute() {
     return GatewayRoutePaths.gatewayPage;
+  }
+
+  static get apiKeysPageRoute() {
+    return GatewayRoutePaths.apiKeysPage;
   }
 
   static get createEndpointPageRoute() {
